rollback spn method to network package
