+++ starport/cmd/network_chain_join.go
@@ -12,6 +12,7 @@ import (
 	"github.com/tendermint/starport/starport/pkg/clispinner"
 	"github.com/tendermint/starport/starport/pkg/cosmosutil"
 	"github.com/tendermint/starport/starport/pkg/xchisel"
+	"github.com/tendermint/starport/starport/services/network"
 )
 
 const (
@@ -57,7 +58,7 @@ func networkChainJoinHandler(cmd *cobra.Command, args []string) error {
 	}
 
 	// get the chain home path
-	home, err := cosmosutil.ChainHome(launchID)
+	home, err := network.ChainHome(launchID)
 	if err != nil {
 		return err
 	}
